Using SwingWorker, Looks for input when in focus
Now uses SwingWorker instead of normal thread to ensure usage of EDT
(Event Dispatch Thread) for events. Previously there was a high chance
that the controls would be unresponsive after compilation. All other
aspects were tested and the problem was narrowed down to the thread
execution. After using the EDT the ship is now responsive every time.
Also, the input map was using the default value for when component in
focus only, now changed to when the component is in the window with
focus